# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Ascend API
  version: 2.0.0
  description: |
    Ascend modular-monolith REST API. All business endpoints require
    Bearer authentication (JWT issued by POST /api/identity/login).

    ## Authentication
    Short-lived access tokens (15 min) are issued at login. A refresh token
    (7 days) allows silent renewal via POST /api/identity/refresh. Tokens are
    also set as httpOnly cookies for browser clients.

    ## Multi-tenancy
    Every JWT carries a `tenantId` claim. The backend enforces tenant isolation
    via application-layer filtering and Row-Level Security. Never pass
    `tenantId` as a request parameter.

    ## Money
    All money values are **integer cents** (e.g. $10.99 → 1099).

    ## Timestamps
    All timestamps are **Unix epoch milliseconds** (integer).

    ## IDs
    UUID v7 strings with a resource-type prefix (ord_, prod_, cust_, etc.).

    ## Error envelope
    All errors:
    ```json
    { "error": { "code": "string", "message": "string", "requestId": "string" } }
    ```

    ## Pagination
    List endpoints use cursor-based pagination unless noted.
    Pass `cursor` from `nextCursor` in the previous response.
servers:
  - url: /
    description: Ascend backend

# ── Security ──────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string, example: unauthenticated }
            message: { type: string }
            requestId: { type: string }
...
    TokenPair:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        expiresIn: { type: integer, example: 900 }
    Role:
      type: string
      enum: [owner, manager, cashier]
...
    OrderLine:
      type: object
      properties:
        productId: { type: string }
        qty: { type: integer }
        unit_price_cents: { type: integer }
        line_total_cents: { type: integer }
        discountCents: { type: integer }
    Order:
      type: object
      properties:
        id: { type: string }
        status: { type: string, enum: [open, completed, refunded, voided] }
        totalCents: { type: integer }
        discountCents: { type: integer }
        customerId: { type: string }
        stateCode: { type: string }
        lines: { type: array, items: { $ref: "#/components/schemas/OrderLine" } }
        created_at: { type: integer }

    # ── Customers ──────────────────────────────────────────────────────────────
...
    ProductionOrder:
      type: object
      properties:
        id: { type: string }
        product_id: { type: string }
        quantity: { type: integer }
        status: { type: string, enum: [draft, in_progress, completed] }
        created_at: { type: integer }
...
  responses:
...
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Conflict:
      description: Business rule violation.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
security:
  - bearerAuth: []

# ═══════════════════════════════════════════════════════════════════════════════
paths:
...
  /api/v1/orders:
    get:
      summary: List orders
      operationId: listOrders
      tags: [Orders]
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [open, completed, refunded, voided] }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: cursor
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Order" }
    post:
      summary: Create order
      operationId: createOrder
      tags: [Orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [lines]
              properties:
                lines:
                  type: array
                  items: { $ref: "#/components/schemas/OrderLine" }
                discountCents: { type: integer }
                customerId: { type: string }
                stateCode: { type: string }
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
  /api/v1/orders/{id}:
    get:
      summary: Get order
      operationId: getOrder
      tags: [Orders]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/orders/{id}/refund:
    post:
      summary: Refund order
      operationId: refundOrder
      tags: [Orders]
      description: Manager or above.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                lineIds: { type: array, items: { type: string } }
                method: { type: string, enum: [original, store_credit] }
      responses:
        "200": { description: Refunded }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/orders/{id}/void:
    post:
      summary: Void order
      operationId: voidOrder
      tags: [Orders]
      description: Manager or above. Same-day only.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: Voided }
        "409": { $ref: "#/components/responses/Conflict" }

  # ── Payments ────────────────────────────────────────────────────────────────
...
  /api/v1/manufacturing/orders:
    get:
      summary: List production orders
      operationId: listProductionOrders
      tags: [Manufacturing]
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [draft, in_progress, completed] }
      responses:
        "200":
          description: Production orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/ProductionOrder" }
    post:
      summary: Create production order with BOM
      operationId: createProductionOrder
      tags: [Manufacturing]
      description: Manager or above.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [product_id, quantity, bomLines]
              properties:
                product_id: { type: string }
                quantity: { type: integer }
                bomLines:
                  type: array
                  items:
                    type: object
                    required: [material_id, qty_required]
                    properties:
                      material_id: { type: string }
                      qty_required: { type: number }
      responses:
        "201":
          description: Production order created with BOM lines
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProductionOrder" }
...
  /api/v1/purchasing/orders:
    get:
      summary: List purchase orders
      operationId: listPurchaseOrders
      tags: [Purchasing]
      responses:
        "200": { description: Purchase orders }
    post:
      summary: Create purchase order
      operationId: createPurchaseOrder
      tags: [Purchasing]
      description: Manager or above.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [supplierId, lines]
              properties:
                supplierId: { type: string }
                lines:
                  type: array
                  items:
                    type: object
                    required: [productId, quantity]
                    properties:
                      productId: { type: string }
                      quantity: { type: integer }
                      unitCostCents: { type: integer }
      responses:
        "201": { description: PO created }
...
  /api/v1/sso/config:
    get:
      summary: Get SSO configuration
      operationId: getSsoConfig
      tags: [SSO]
      description: Owner only.
      responses:
        "200": { description: OIDC/SAML config }
    put:
      summary: Upsert SSO configuration
      operationId: setSsoConfig
      tags: [SSO]
      description: Owner only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled, clientId, clientSecret, discoveryUrl]
              properties:
                enabled: { type: boolean }
                providerName: { type: string }
                clientId: { type: string }
                clientSecret: { type: string }
                discoveryUrl: { type: string }
                defaultRole: { $ref: "#/components/schemas/Role" }
      responses:
        "200": { description: SSO config saved }
    delete:
      summary: Delete SSO configuration
      operationId: deleteSsoConfig
      tags: [SSO]
      description: Owner only.
      responses:
        "204": { description: SSO config deleted }
  /api/v1/sso/initiate:
    post:
      summary: Initiate OIDC login
      operationId: ssoInitiate
      security: []
      tags: [SSO]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenantId, redirectUri]
              properties:
                tenantId: { type: string }
                redirectUri: { type: string }
      responses:
        "200": { description: OIDC authorization URL }
  /api/v1/sso/callback:
    post:
      summary: Handle OIDC callback
      operationId: ssoCallback
      security: []
      tags: [SSO]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [state, code, redirectUri]
              properties:
                state: { type: string }
                code: { type: string }
                redirectUri: { type: string }
      responses:
        "200":
          description: Token pair for the authenticated user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenPair" }